grpc-client/requirements.txt
aniso8601==9.0.1
attrs==23.1.0
blinker==1.7.0
click==8.1.7
Flask==2.3.3
flask-restx==1.2.0
grpcio==1.60.0
grpcio-tools==1.60.0
importlib-metadata==7.0.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.3
protobuf==4.25.1
pyrsistent==0.20.0
pytz==2023.3.post1
Werkzeug==2.3.8
zipp==3.17.0
